+++ src/Symfony/Contracts/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Add a `$priority` argument to the `Required` attribute to control the order in which required methods are called
  * Add support for the `trailers` info in `ResponseInterface`
+ * Add `ItemInterface::METADATA_SAVE_FAILED` to report in the `$metadata` argument of `CacheInterface::get()` that the computed value could not be saved
 
 3.7
 ---